+++ crates/forge/src/running.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use blockifier::execution::entry_point::{
     CallEntryPoint, CallType, EntryPointExecutionContext, ExecutionResources,
 };
@@ -25,8 +25,11 @@ use cheatnet::constants as cheatnet_constants;
 use cheatnet::forking::state::ForkStateReader;
 use cheatnet::state::{CheatnetState, ExtendedStateReader};
 use conversions::StarknetConversions;
-use starknet::core::types::{BlockId, BlockTag};
+use starknet::core::types::BlockTag::Latest;
+use starknet::core::types::{BlockId, BlockTag, MaybePendingBlockWithTxHashes};
 use starknet::core::utils::get_selector_from_name;
+use starknet::providers::jsonrpc::HttpTransport;
+use starknet::providers::{JsonRpcClient, Provider};
 use starknet_api::core::PatriciaKey;
 use starknet_api::core::{ContractAddress, EntryPointSelector};
 use starknet_api::deprecated_contract_class::EntryPointType;
@@ -34,10 +37,17 @@ use starknet_api::hash::StarkHash;
 use starknet_api::patricia_key;
 use starknet_api::transaction::Calldata;
 use test_collector::{ForkConfig, TestCase};
+use tokio::runtime::Runtime;
+use tokio::sync::mpsc::Sender;
+use url::Url;
 
-use crate::scarb::{ForkTarget, StarknetContractArtifacts};
+use crate::scarb::ForkTarget;
 use crate::test_case_summary::TestCaseSummary;
-use crate::test_execution_syscall_handler::{TestExecutionState, TestExecutionSyscallHandler};
+
+use crate::test_execution_syscall_handler::TestExecutionSyscallHandler;
+use crate::{RunnerConfig, RunnerParams};
+
+use crate::test_execution_syscall_handler::TestExecutionState;
 
 /// Builds `hints_dict` required in `cairo_vm::types::program::Program` from instructions.
 fn build_hints_dict<'b>(
@@ -63,6 +73,27 @@ fn build_hints_dict<'b>(
         hint_offset += instruction.body.op_size();
     }
     (hints_dict, string_to_hint)
+}
+
+pub(crate) async fn blocking_run_from_test(
+    args: Vec<Felt252>,
+    case: Arc<TestCase>,
+    runner: Arc<SierraCasmRunner>,
+    runner_config: Arc<RunnerConfig>,
+    runner_params: Arc<RunnerParams>,
+    sender: Option<Sender<()>>,
+) -> Result<TestCaseSummary> {
+    tokio::task::spawn_blocking(move || {
+        run_test_case(
+            args,
+            &case,
+            &runner,
+            &runner_config,
+            &runner_params,
+            &sender,
+        )
+    })
+    .await?
 }
 
 fn build_context() -> EntryPointExecutionContext {
@@ -115,14 +146,12 @@ fn build_syscall_handler<'a>(
 
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn run_test_case(
-    package_root: &Utf8PathBuf,
-    runner: &SierraCasmRunner,
-    case: &TestCase,
-    fork_targets: &[ForkTarget],
-    contracts: &HashMap<String, StarknetContractArtifacts>,
-    predeployed_contracts: &Utf8PathBuf,
     args: Vec<Felt252>,
-    environment_variables: &HashMap<String, String>,
+    case: &TestCase,
+    runner: &SierraCasmRunner,
+    runner_config: &Arc<RunnerConfig>,
+    runner_params: &Arc<RunnerParams>,
+    _sender: &Option<Sender<()>>,
 ) -> Result<TestCaseSummary> {
     let available_gas = if let Some(available_gas) = &case.available_gas {
         Some(*available_gas)
@@ -143,8 +172,14 @@ pub(crate) fn run_test_case(
     let (hints_dict, string_to_hint) = build_hints_dict(instructions.clone());
 
     let state_reader = ExtendedStateReader {
-        dict_state_reader: cheatnet_constants::build_testing_state(predeployed_contracts),
-        fork_state_reader: get_fork_state_reader(package_root, fork_targets, &case.fork_config),
+        dict_state_reader: cheatnet_constants::build_testing_state(
+            &runner_params.predeployed_contracts,
+        ),
+        fork_state_reader: get_fork_state_reader(
+            &runner_config.workspace_root,
+            &runner_config.fork_targets,
+            &case.fork_config,
+        )?,
     };
     let mut context = build_context();
     let mut execution_resources = ExecutionResources::default();
@@ -161,8 +196,8 @@ pub(crate) fn run_test_case(
         CheatableSyscallHandler::new(syscall_handler, &mut cheatnet_state);
 
     let mut test_execution_state = TestExecutionState {
-        environment_variables,
-        contracts,
+        environment_variables: &runner_params.environment_variables,
+        contracts: &runner_params.contracts,
     };
     let mut test_execution_syscall_handler = TestExecutionSyscallHandler::new(
         cheatable_syscall_handler,
@@ -188,6 +223,7 @@ pub(crate) fn run_test_case(
                 error.to_string().replace(" Custom Hint Error: ", "\n    ")
             )),
             arguments: args,
+            fuzzing_statistic: None,
         }),
 
         Err(err) => Err(err.into()),
@@ -195,51 +231,71 @@ pub(crate) fn run_test_case(
 }
 
 fn get_fork_state_reader(
-    package_root: &Utf8PathBuf,
+    workspace_root: &Utf8PathBuf,
     fork_targets: &[ForkTarget],
     fork_config: &Option<ForkConfig>,
-) -> Option<ForkStateReader> {
+) -> Result<Option<ForkStateReader>> {
     match &fork_config {
-        Some(ForkConfig::Params(url, block_id)) => Some(ForkStateReader::new(
-            url,
-            *block_id,
-            Some(package_root.join(".snfoundry_cache").as_ref()),
-        )),
-        Some(ForkConfig::Id(name)) => {
-            find_params_and_build_fork_state_reader(package_root, fork_targets, name)
+        Some(ForkConfig::Params(url, mut block_id)) => {
+            if let BlockId::Tag(Latest) = block_id {
+                block_id = get_latest_block_number(url)?;
+            }
+            Ok(Some(ForkStateReader::new(
+                url,
+                block_id,
+                Some(workspace_root.join(".snfoundry_cache").as_ref()),
+            )))
         }
-        _ => None,
+        Some(ForkConfig::Id(name)) => {
+            find_params_and_build_fork_state_reader(workspace_root, fork_targets, name)
+        }
+        _ => Ok(None),
+    }
+}
+
+fn get_latest_block_number(url: &str) -> Result<BlockId> {
+    let client = JsonRpcClient::new(HttpTransport::new(Url::parse(url).unwrap()));
+    let runtime = Runtime::new().expect("Could not instantiate Runtime");
+
+    match runtime.block_on(client.get_block_with_tx_hashes(BlockId::Tag(Latest))) {
+        Ok(MaybePendingBlockWithTxHashes::Block(block)) => Ok(BlockId::Number(block.block_number)),
+        _ => Err(anyhow!("Could not get the latest block number".to_string())),
     }
 }
 
 fn find_params_and_build_fork_state_reader(
-    package_root: &Utf8PathBuf,
+    workspace_root: &Utf8PathBuf,
     fork_targets: &[ForkTarget],
     fork_alias: &str,
-) -> Option<ForkStateReader> {
-    let fork = fork_targets.iter().find(|fork| fork.name == fork_alias);
-
-    let block_id = fork?
-        .block_id
-        .iter()
-        .map(|(id_type, value)| match id_type.as_str() {
-            "number" => Some(BlockId::Number(value.parse().unwrap())),
-            "hash" => Some(BlockId::Hash(value.to_field_element())),
-            "tag" => match value.as_str() {
-                "Latest" => Some(BlockId::Tag(BlockTag::Latest)),
-                "Pending" => Some(BlockId::Tag(BlockTag::Pending)),
+) -> Result<Option<ForkStateReader>> {
+    if let Some(fork) = fork_targets.iter().find(|fork| fork.name == fork_alias) {
+        let block_id = fork
+            .block_id
+            .iter()
+            .map(|(id_type, value)| match id_type.as_str() {
+                "number" => Some(BlockId::Number(value.parse().unwrap())),
+                "hash" => Some(BlockId::Hash(value.to_field_element())),
+                "tag" => match value.as_str() {
+                    "Latest" => Some(BlockId::Tag(BlockTag::Latest)),
+                    "Pending" => Some(BlockId::Tag(BlockTag::Pending)),
+                    _ => unreachable!(),
+                },
                 _ => unreachable!(),
-            },
-            _ => unreachable!(),
-        })
-        .collect::<Vec<_>>();
-    let [Some(block_id)] = block_id[..] else {
-        return None;
-    };
+            })
+            .collect::<Vec<_>>();
+        let [Some(mut block_id)] = block_id[..] else {
+            return Ok(None);
+        };
 
-    Some(ForkStateReader::new(
-        &fork?.url,
-        block_id,
-        Some(package_root.join(".snfoundry_cache").as_ref()),
-    ))
+        if let BlockId::Tag(Latest) = block_id {
+            block_id = get_latest_block_number(&fork.url)?;
+        }
+        return Ok(Some(ForkStateReader::new(
+            &fork.url,
+            block_id,
+            Some(workspace_root.join(".snfoundry_cache").as_ref()),
+        )));
+    }
+
+    Ok(None)
 }